Throw error if where condition is empty
FossilOrigin-Name: 5fe9a5dbeecc9519203f3c7ae660da376097d477bb37c4af888af615cb4bc486

+++ src/include/lib/Paheko/UserTemplate/Functions.php
@@ -409,6 +409,10 @@ class Functions
 			}
 		}
 
+		if (!count($where)) {
+			throw new Brindille_Exception('Missing parameters for delete');
+		}
+
 		$where = implode(' AND ', $where);
 		$db->delete($table, $where, $args);
 	}
